download_emails: move to command class
